requirements.txt
email-validator>=2.2.0
flask-cors>=5.0.1
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
pyjwt>=2.10.1
pymongo>=4.12.0
bcrypt>=4.3.0
numpy<2.0.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
datasets>=2.14.0
huggingface_hub>=0.19.0
keras>=3.0.0
jax>=0.4.20
jaxlib>=0.4.20
opencv-python>=4.5.0
mediapipe>=0.10.0
tensorflow>=2.15.0

Pillow>=9.0.0

# Optional ML libraries (uncomment when needed)